Extract TimelineTaskCardContent and AllDayTaskCard shared components
Day view task cards and all-day chips now render the exact same components
as multi view, making visual divergence impossible:

- TimelineTaskCardContent.jsx: extracted from TimeGrid lines 511-831;
  includes all three layouts (imported/narrow/wide), tags, notes button,
  action buttons, inline editing, AI subtasks, notes panels
- AllDayTaskCard.jsx: extracted from CalendarHeader all-day card;
  includes tablet touch handlers, inline editing, notes/subtask panel,
  project badge, full/compact action button layouts

TimeGrid, DayView, DayViewAllDaySection, CalendarHeader all updated to
use these shared components. DayViewAllDaySection switches from the
custom AllDayChip/GroupChips overflow approach to the same stacked
space-y-1 layout used by multi view.
